crawler_flink_kafka/farmacia/requirements.txt
kafka-python==0.9.4
scrapy==1.5.0